Version 1.3.0 - Local shell emulation
This change removes the built-in telnet capability in favor of allowing
the user to run an arbitrary shell command, essentially turning the
DMD5620 emulator into a local terminal emulator.

The options to specify host and port have been removed. In their place,
an option to supply a shell command to execute has been added. If it
is not supplied, the user's default login shell will be used.

A second option has been added to allow the user to specify whether
backspace generates ^H (the default), or ^? (DEL).

A rudimentary manpage has been added.