Limit the number of steps per loop
It was possible for the code to attempt to run up to 10,000,000
simulated steps each time through the main simulation loop. This was
much too many simulated steps, and could result in starving the other
GTK threads of resources.

This change drops that maximum value way way down to 150,000 steps. This
should guarantee that the GTK tick callback can always complete in a
timely fashion and not steal resources from the other threads, while
still giving excellent CPU performance.